+++ components/combobox/combobox.jsx
@@ -35,7 +35,7 @@ import KeyBuffer from '../../utilities/key-buffer';
 import keyLetterMenuItemSelect from '../../utilities/key-letter-menu-item-select';
 import mapKeyEventCallbacks from '../../utilities/key-callbacks';
 import menuItemSelectScroll from '../../utilities/menu-item-select-scroll';
-import { DIRECTIONS } from '../utilities/direction/direction-settings';
+import { DIRECTIONS } from '../utilities/private/direction/direction-settings';
 
 import checkProps from './check-props';
