+++ components/component-docs.json
@@ -3763,6 +3763,27 @@
                 "required": false,
                 "description": "CSS classes to be added to tag with `.slds-dropdown__header`. Uses `classNames` [API](https://github.com/JedWatson/classnames)."
             },
+            "direction": {
+                "type": {
+                    "name": "enum",
+                    "value": [
+                        {
+                            "value": "DIRECTIONS.LTR",
+                            "computed": true
+                        },
+                        {
+                            "value": "DIRECTIONS.RTL",
+                            "computed": true
+                        }
+                    ]
+                },
+                "required": false,
+                "description": "Establishes directional context for component. Defaults to left-to-right.",
+                "defaultValue": {
+                    "value": "DIRECTIONS.LTR",
+                    "computed": true
+                }
+            },
             "events": {
                 "type": {
                     "name": "shape",
@@ -7303,7 +7324,7 @@
                                 ]
                             },
                             "required": true,
-                            "description": "Region wrap children in styling specific to that region."
+                            "description": "Region wrap children in styling specific to that region. When `tertiary`\nregion is used, secondary region only supports four list items."
                         }
                     },
                     "name": "region",
